+++ graphiant_sdk/configuration.py
@@ -554,7 +554,7 @@ conf = graphiant_sdk.Configuration(
                "OS: {env}\n"\
                "Python Version: {pyversion}\n"\
                "Version of the API: 1.0.0\n"\
-               "SDK Package Version: 26.3.2".\
+               "SDK Package Version: 26.3.3".\
                format(env=sys.platform, pyversion=sys.version)
 
     def get_host_settings(self) -> List[HostSetting]: